fix(koswat/__main__.py): Renamed main file to fit to python main expected module
