+++ koswat/__init__.py
@@ -1,11 +1,1 @@
 __version__ = "0.11.0"
-
-import logging
-
-import koswat.main
-
-# if __name__ == "__main__":
-#     try:
-#         run_analysis()
-#     except Exception as e_info:
-#         logging.error(str(e_info))
